requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.2
python-multipart==0.0.7
filelock>=3.12
python-dotenv>=1.0

numpy>=1.26
scikit-learn>=1.4
yt-dlp>=2026.3.17
imageio>=2.37
imageio-ffmpeg>=0.6

# Dev / testing
httpx>=0.27
pytest>=8
pytest-asyncio>=0.23
